requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
joblib==1.3.2
pandas==2.1.4
httpx==0.28.1
alembic==1.11.1
lightgbm==4.1.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
fastapi-mail==1.4.1